+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Adds the wildcard token, denoted as `*`, that matches anything.
+
 ## 0.1.1
 
 * Lowers SDK constraint so that Pub can analyze the package and generate